formatting

+++ libft/ft_range.c
@@ -6,21 +6,13 @@
 /*   By: kmaitski <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/20 14:41:12 by kmaitski          #+#    #+#             */
-/*   Updated: 2017/04/20 15:53:15 by kmaitski         ###   ########.fr       */
+/*   Updated: 2017/05/26 14:25:14 by kmaitski         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
-/* 
- * ===  FUNCTION  ==============================================================
- *         Name:  get_range
- *  Description:  Returns the range for ft_range to allocate enough memory using 
- *  			  malloc.
- * =============================================================================
- */
-
-static unsigned int	get_range (int start, int end)
+static unsigned int	get_range(int start, int end)
 {
 	unsigned int	range;
 
@@ -29,27 +21,18 @@ static unsigned int	get_range (int start, int end)
 		range = start - end + 1;
 	else
 		range = end - start + 1;
-	return ((unsigned)range);
-}		/* -----  end of function get_range  ----- */
+	return (range);
+}
 
-/* 
- * ===  FUNCTION  ===============================================================
- *         Name:  ft_range
- *  Description:  It must allocate (with malloc()) an array of integers, fill it
- *  			  with consecutive values that begin at start and end at end 
- *  			  (Including start and end !), then return a pointer to the first
- *  			  value of the array.
- * ==============================================================================
- */
-int					*ft_range (int start, int end)
+int					*ft_range(int start, int end)
 {
 	unsigned int	range;
 	int				*array;
 	unsigned int	i;
-	
+
 	i = 0;
 	range = get_range(start, end);
-	if(!(array = (int *)malloc(sizeof(int) * range)))
+	if (!(array = (int *)malloc(sizeof(int) * range)))
 		return (NULL);
 	if (range == 1)
 		array[0] = start;
@@ -60,4 +43,4 @@ int					*ft_range (int start, int end)
 		while (i < range)
 			array[i++] = start++;
 	return (array);
-}		/* -----  end of function ft_range  ----- */
+}
